requirements.txt
amqp==2.5.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
Django==2.2.5
djangorestframework==3.10.3
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.4
lxml==4.4.1
more-itertools==7.2.0
parsel==1.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pytz==2019.2
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
Twisted==19.7.0
urllib3==1.25.6
vine==1.3.0
w3lib==1.21.0
zipp==0.6.0
zope.interface==4.6.0
